requirements.txt
PyJWT
httpx
fastapi
uvicorn